+++ .gitignore
@@ -3,3 +3,5 @@
 local.properties
 /.idea
 /gradle
+/build
+libs/
