--- app/src/main/java/com/ozh/bunchdecorator/example/ShowCaseActivity.kt
@@ -1,24 +1,0 @@
-package com.ozh.bunchdecorator.example
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_show_case.*
-
-class ShowCaseActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_show_case)
-        initViews()
-    }
-
-    private fun initViews() {
-        linear_decor_btn.setOnClickListener {
-            startActivity(Intent(this, LinearDecoratorActivity::class.java))
-        }
-
-        drawable_decor_btn.setOnClickListener {
-            startActivity(Intent(this, CarouselDecoratorActivity::class.java))
-        }
-    }
-}
