+++ app/src/main/java/com/ozh/dd/example/sample/DividerRule.kt
@@ -1,0 +1,10 @@
+package com.ozh.dd.example.sample
+
+import androidx.annotation.IntDef
+
+@Retention(AnnotationRetention.SOURCE)
+@IntDef(
+    Rules.MIDDLE,
+    Rules.END
+)
+annotation class DividerRule
