+++ app/src/main/res/layout/activity_main.xml
@@ -11,5 +11,5 @@
             android:id="@+id/recycler_view"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".LinearDecoratorActivity"/>
+            tools:context="com.ozh.dd.example.LinearDecoratorActivity"/>
 </LinearLayout>
