+++ app/src/main/res/layout/activity_show_case.xml
@@ -7,7 +7,7 @@
         android:layout_height="match_parent"
         android:gravity="center"
         android:orientation="vertical"
-        tools:context=".ShowCaseActivity">
+        tools:context="com.ozh.dd.example.ShowCaseActivity">
 
     <Button
             android:id="@+id/linear_decor_btn"
