+++ lib/src/main/java/com/ozh/dd/Consts.kt
@@ -1,0 +1,3 @@
+package com.ozh.dd
+
+const val UNDEFINE_VIEW_TYPE = -1
