+++ lib/src/main/java/com/ozh/dd/draw/OffsetDrawer.kt
@@ -1,0 +1,14 @@
+package com.ozh.dd.draw
+
+import android.graphics.Rect
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+abstract class OffsetDrawer {
+    abstract fun getItemOffsets(
+        outRect: Rect,
+        childView: View,
+        recyclerView: RecyclerView,
+        state: RecyclerView.State
+    )
+}
