+++ lib/src/main/java/com/ozh/dd/round/RoundMode.kt
@@ -1,0 +1,11 @@
+package com.ozh.dd.round
+
+/**
+ * Enum для описания закругления краев View
+ */
+enum class RoundMode {
+    TOP,
+    BOTTOM,
+    ALL,
+    NONE
+}
