--- lib/src/main/java/com/ozh/dsl/decorators/Consts.kt
@@ -1,3 +1,0 @@
-package com.ozh.dsl.decorators
-
-const val UNDEFINE_VIEW_HOLDER: Int = -1
