--- lib/src/main/java/com/ozh/dsl/decorators/deprecated/LinearItemsDecoration.kt
@@ -1,33 +1,0 @@
-package com.ozh.dsl.decorators.deprecated
-
-import android.graphics.Canvas
-import android.graphics.Rect
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-@Deprecated("Use com.ozh.dsl.decorators.layers.Decorator")
-class LinearItemsDecoration(var drawOver: Boolean = false) : RecyclerView.ItemDecoration() {
-
-    private val drawManager = DrawManager()
-
-    override fun getItemOffsets(outRect: Rect, childView: View, recyclerView: RecyclerView, state: RecyclerView.State) {
-        super.getItemOffsets(outRect, childView, recyclerView, state)
-        if (drawOver.not()) {
-            drawManager.getItemOffsets(outRect, childView, recyclerView, state)
-        }
-    }
-
-    override fun onDraw(canvas: Canvas, recyclerView: RecyclerView, state: RecyclerView.State) {
-        super.onDraw(canvas, recyclerView, state)
-        if(drawOver.not()) {
-            drawManager.onDraw(canvas, recyclerView)
-        }
-    }
-
-    override fun onDrawOver(canvas: Canvas, recyclerView: RecyclerView, state: RecyclerView.State) {
-        super.onDrawOver(canvas, recyclerView, state)
-        if (drawOver) {
-            drawManager.onDrawOver(canvas, recyclerView)
-        }
-    }
-}
