--- lib/src/main/java/com/ozh/dsl/decorators/layers/Layer.kt
@@ -1,6 +1,0 @@
-package com.ozh.dsl.decorators.layers
-
-class Layer(
-    val viewItemType: Int,
-    val layerDrawer: LayerDrawer
-)
