--- lib/src/main/java/com/ozh/dsl/decorators/layers/Offset.kt
@@ -1,6 +1,0 @@
-package com.ozh.dsl.decorators.layers
-
-class Offset(
-    val viewItemType: Int,
-    val offsetDrawer: OffsetDrawer
-)
