--- lib/src/main/java/com/ozh/dsl/decorators/layers/OffsetDrawer.kt
@@ -1,14 +1,0 @@
-package com.ozh.dsl.decorators.layers
-
-import android.graphics.Rect
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-abstract class OffsetDrawer {
-    abstract fun getItemOffsets(
-        outRect: Rect,
-        childView: View,
-        recyclerView: RecyclerView,
-        state: RecyclerView.State
-    )
-}
