--- lib/src/main/java/com/ozh/dsl/decorators/round/RoundMode.kt
@@ -1,11 +1,0 @@
-package com.ozh.dsl.decorators.round
-
-/**
- * Enum для описания закругления краев View
- */
-enum class RoundMode {
-    TOP,
-    BOTTOM,
-    ALL,
-    NONE
-}
