--- lib/src/main/java/com/ozh/dsl/decorators/sample/DividerRule.kt
@@ -1,10 +1,0 @@
-package com.ozh.dsl.decorators.sample
-
-import androidx.annotation.IntDef
-
-@Retention(AnnotationRetention.SOURCE)
-@IntDef(
-    Rules.MIDDLE,
-    Rules.END
-)
-annotation class DividerRule
